extract_mask_from_images_using_ml/requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
bleach==1.5.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.2
Django==3.0.8
Flask==1.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==0.9999999
image==1.5.32
imageio==2.9.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.2.0
klepto==0.1.9
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mrcnn==0.2
networkx==2.4
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
pox==0.2.8
protobuf==3.12.2
pykakasi==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
tensorboard==1.9.0
tensorflow==1.5.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tifffile==2020.7.24
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
